Medium Update (UNTESTED)
Added the code for the weapon buildings
This version WILL CRASH, unless you add the images manually. This is
because I haven't made the graphics for the buildings yet so the game
tries to load something that doesn't exist resulting in a crash.
I also haven't tested if everything is working correctly.